python-sdk-resource-creation-samples/create-database/requirements.txt
azure-mgmt-resource
azure-mgmt-rdbms
azure-identity
mysql
mysql-connector